feature: Webpack Spring Cleaning
This cleans up the webpack configurations by deduplicating a lot of the
shared configurations.
