--- webpack/envs/index.js
@@ -1,17 +1,0 @@
-// @ts-check
-
-const { clientCommonConfig } = require("./clientCommonConfig")
-const { clientDevelopmentConfig } = require("./clientDevelopmentConfig")
-const { clientProductionConfig } = require("./clientProductionConfig")
-const { novoDevelopmentConfig } = require("./novoDevelopmentConfig")
-const { novoProductionConfig } = require("./novoProductionConfig")
-const { serverConfig } = require("./serverConfig")
-
-export {
-  clientCommonConfig,
-  clientDevelopmentConfig,
-  clientProductionConfig,
-  novoDevelopmentConfig,
-  novoProductionConfig,
-  serverConfig,
-}
