--- webpack/plugins/index.js
@@ -1,5 +1,0 @@
-// @ts-check
-
-const { bundleAnalyzer } = require("./bundleAnalyzer")
-
-export { bundleAnalyzer }
